+++ Changelog.rst
@@ -3,6 +3,10 @@ Version NEXTVERSION
 
 **2026-??-??**
 
+* New default backend for netCDF-4 in `cf.write`: ``h5netcdf-h5py``,
+  that allows control of the internal file metadata via the new
+  ``h5py_options`` parameter
+  (https://github.com/NCAS-CMS/cf-python/issues/924)
 * New default backend for netCDF-4 in `cf.read` that allows parallel
   reading: ``h5netcdf-pyfive``
   (https://github.com/NCAS-CMS/cf-python/issues/912)
@@ -12,8 +16,6 @@ Version NEXTVERSION
 * Changed dependency: ``cfdm>=1.13.1.0, <1.13.2.0``
 * Fix for subspacing with cyclic `cf.wi` and `cf.wo` arguments
   (https://github.com/NCAS-CMS/cf-python/issues/887)
-
-----
 
 Version 3.19.0
 --------------